Moved ng-data-context to /Generic First compilation of generic skip chat component
