+++ packages/Angular/Generic/data-context/src/lib/ng-data-context.component.ts
