+++ packages/Angular/Generic/skip-chat/package.json
@@ -24,6 +24,7 @@
   },
   "dependencies": {
     "tslib": "^2.3.0",
+    "@angular/forms": "18.0.2",
     "@memberjunction/core-entities": "2.13.4",
     "@memberjunction/core": "2.13.4",
     "@memberjunction/global": "2.13.4",
@@ -34,7 +35,6 @@
     "angular-plotly.js": "^5.2.2",
     "plotly.js-dist-min": "^2.3.4",
     "@types/plotly.js-dist-min": "^2.3.4",
-    "ngx-markdown": "18.0.0",
     "@angular/cdk": "18.0.2"
   },
   "sideEffects": false
