+++ packages/Angular/Generic/skip-chat/src/lib/dynamic-report/linear-report.css
