+++ packages/Angular/Generic/skip-chat/src/lib/dynamic-report/linear-report.html
@@ -1,0 +1,62 @@
+@if(SkipData) {
+    <button kendoButton *ngIf="ShowCreateReportButton && !matchingReportID" 
+                        (click)="askCreateReport()" 
+                        [disabled]="isCreatingReport"
+                        class="create-report-button">Create Report</button>
+    <div *ngIf="matchingReportID!==null" class="report-link" (click)="clickMatchingReport()">Report: {{matchingReportName}} (ID: {{matchingReportID}}) Created From This Message</div>
+    <div *ngIf="SkipData?.userExplanation" class="report-tab-title">
+        <h3>Explanation</h3>
+        <div>
+            {{SkipData.userExplanation || 'No Explanation Provided'}}
+        </div>
+    </div>
+    <kendo-expansionpanel
+        *ngIf="IsChart"
+        [title]="'Chart'"
+        [expanded]="ExpandAll"
+        >
+        <skip-dynamic-chart 
+            *ngIf="IsChart" 
+            #theChart 
+            [AutoResizeChart]="false"
+            [SkipData]="SkipData" 
+            style="display:block;">
+        </skip-dynamic-chart>
+    </kendo-expansionpanel>                     
+    <kendo-expansionpanel
+        *ngIf="!IsChart"
+        [title]="'Table'"
+        [expanded]="ExpandAll"
+        mjSkipResize="true"
+    >
+        <skip-dynamic-grid *ngIf="!IsChart" 
+                        [GridHeight]="300"
+                        #theGrid [SkipData]="SkipData"
+                        ></skip-dynamic-grid>  
+    </kendo-expansionpanel>
+    <kendo-expansionpanel
+        *ngIf="SkipData?.analysis" 
+        [title]="'Analysis'"
+        [expanded]="ExpandAll"
+        >
+        <!-- <markdown [data]="SkipData.analysis"></markdown> -->
+    </kendo-expansionpanel>                     
+
+  <kendo-dialog
+    title="Please confirm"
+    *ngIf="confirmCreateReportDialogOpen"
+    (close)="closeCreateReport('no')"
+    [minWidth]="250"
+    [width]="450"
+  >
+  <p style="margin: 30px; text-align: center;">
+    Would you like to create a new report from this conversation element? If you choose to continue you'll be notified when the report has been created.
+  </p>
+  <kendo-dialog-actions>
+    <button kendoButton (click)="closeCreateReport('no')">No</button>
+    <button kendoButton (click)="closeCreateReport('yes')" themeColor="primary">
+      Yes
+    </button>
+  </kendo-dialog-actions>
+  </kendo-dialog>
+}
