+++ packages/Angular/Generic/skip-chat/src/lib/dynamic-report/linear-report.ts
@@ -6,34 +6,24 @@ import { LogError, Metadata, RunView } from '@memberjunction/core';
 import { ReportEntity } from '@memberjunction/core-entities';
 import { DataContext } from '@memberjunction/data-context';
 import { GraphQLDataProvider } from '@memberjunction/graphql-dataprovider';
-import { DrillDownInfo, SkipDynamicReportDrillDownComponent } from './dynamic-drill-down';
-import { MJTabStripComponent, TabEvent } from '@memberjunction/ng-tabstrip';
 
-// This component is used for dynamically rendering report data, it is wrapped by app-single-report which gets
-// info from the database. This can also be used directly to render a dynamic report that is NOT saved in the DB
-// which is what Skip does in its conversational UI
-
+/**
+ * This component is used for dynamically rendering Skip Reports
+ */
 @Component({
-  selector: 'skip-dynamic-report',
-  styleUrls: ['./dynamic-report.css'],
-  templateUrl: './dynamic-report.html',
+  selector: 'skip-dynamic-linear-report',
+  styleUrls: ['./linear-report.css'],
+  templateUrl: './linear-report.html',
 })
-export class SkipDynamicReportComponent implements AfterViewInit, AfterViewChecked {
-  @Input() ShowDetailsTab: boolean = false;
+export class SkipDynamicLinearReportComponent implements AfterViewInit {
   @Input() ShowCreateReportButton: boolean = false;
   @Input() ConversationID: string | null = null;
   @Input() ConversationName: string | null = null;
   @Input() ConversationDetailID: string | null = null;
   @Input() DataContext!: DataContext;
   @Input() ReportEntity?: ReportEntity;
-  @Input() LayoutMode: 'linear' | 'tabs' = 'tabs';
-  @Input() LinearExpandAll: boolean = true;
+  @Input() ExpandAll: boolean = true;
   @Input() SkipData: SkipAPIAnalysisCompleteResponse | undefined;
-  @Input() AllowDrillDown: boolean = true;
-
-  public DrillDowns: DrillDownInfo[] = [];
-
-  @ViewChild('drillDownComponent', { static: false }) drillDownComponent!: SkipDynamicReportDrillDownComponent;
 
   constructor(
     public sharedService: SharedService,
@@ -51,7 +41,7 @@ export class SkipDynamicReportComponent implements AfterViewInit, AfterViewCheck
       if (this.ShowCreateReportButton) {
         // check to see if a report has been created that is linked to this ConvoID/ConvoDetailID
         // if so don't allow the user to create another report, show a link to the existing one
-        const cachedItem = SkipDynamicReportComponent._reportCache.find(
+        const cachedItem = SkipDynamicLinearReportComponent._reportCache.find(
           (x) => x.conversationId === this.ConversationID && x.conversationDetailId === this.ConversationDetailID
         );
         if (cachedItem) {
@@ -68,7 +58,7 @@ export class SkipDynamicReportComponent implements AfterViewInit, AfterViewCheck
             this.matchingReportID = item.ID;
             this.matchingReportName = item.Name;
             // cache for future to avoid db call
-            SkipDynamicReportComponent._reportCache.push({
+            SkipDynamicLinearReportComponent._reportCache.push({
               reportId: item.ID,
               conversationId: this.ConversationID,
               reportName: item.Name,
@@ -94,47 +84,7 @@ export class SkipDynamicReportComponent implements AfterViewInit, AfterViewCheck
       this.router.navigate(['resource', 'report', this.matchingReportID]);
     }
   }
-
-  public activeTabIndex: number = 0;
-  public onTabSelect(e: TabEvent): void {
-    this.activeTabIndex = e.index;
-    this.sharedService.InvokeManualResize(100); // for the first tab, chart, have a longer delay because that is the plotly chart which takes a moment to render and needs to be resized after
-    if (e.index === 0) {
-      // special case, do an extra resize after a longer delay since plotly charts take a moment to render
-      this.sharedService.InvokeManualResize(750);
-    }
-  }
-
-  public isTabSelected(index: number) {
-    // the index passed in is a value that is the index of the tab at the DESIGN TIME IN THE HTML
-    // which is based on the maximum # of possible tabs that can exist.
-    // If some of the tabs are NOT showing, then we have fewer tabs than the max
-    // for this reason, we need to
-    return this.activeTabIndex === index - this.getCurrentTabOffset(index);
-  }
-
-  protected getCurrentTabOffset(currentTabIndex: number): number {
-    let offset = 0;
-    switch (currentTabIndex) {
-      case 0:
-        // chart tab, no change
-        break;
-      case 1:
-        // table tab. If chart tab isn't showing, then we need to offset by 1
-        if (!this.IsChart) offset++;
-        break;
-      case 2:
-        // drill down tab. If chart tab isn't showing, then we need to offset
-        if (!this.IsChart) offset++;
-        break;
-      default:
-        // rest of the tabs above the first 3 are always showing, so we need to offset by the number of tabs that are not showing
-        if (!this.IsChart) offset++;
-        if (this.DrillDowns.length === 0 || !this.AllowDrillDown) offset++;
-        break;
-    }
-    return offset;
-  }
+ 
 
   public get Columns(): SkipColumnInfo[] {
     return this.SkipData?.tableDataColumns || [];
@@ -271,41 +221,6 @@ export class SkipDynamicReportComponent implements AfterViewInit, AfterViewCheck
     catch (err) {
       this.sharedService.CreateSimpleNotification('Error refreshing report', 'error', 2500);
       console.error(err);
-    }
-  }
-
-  public async handleChartDrillDown(info: DrillDownInfo) {
-    this.handleDrillDown(info);
-  }
-  public handleGridDrillDown(info: DrillDownInfo) {
-    this.handleDrillDown(info);
-  }
-  private _drillDownSelectTab: number = -1;
-  protected handleDrillDown(info: DrillDownInfo) {
-    // first, make sure that the info is not already in the drill down list
-    const idx = this.DrillDowns.findIndex((x) => x.EntityName === info.EntityName && x.Filter === info.Filter);
-    if (idx >= 0) {
-      // here we already have the drill down, so just select the tab
-      this._drillDownSelectTab = idx;
-    } else {
-      // just add to the info, Angular binding will then update the drill down tabs
-      if (this.SkipData?.drillDown?.baseFilter && (!info.BaseFilter || info.BaseFilter.length === 0))
-        info.BaseFilter = this.SkipData?.drillDown?.baseFilter; // add the base filter if we have it from Skip and it wasn't already set
-
-      this.DrillDowns.push(info);
-      // now make sure the drill down tab is selected, but wait a bit to ensure angular has updated the view
-      this._drillDownSelectTab = this.DrillDowns.length - 1;
-    }
-  }
-
-  public ngAfterViewChecked(): void {
-    if (this._drillDownSelectTab >= 0 && this.drillDownComponent) {
-      if (this.IsChart)
-        this.activeTabIndex = 2; // chart tab IS showing show index of drill down is 2
-      else this.activeTabIndex = 1; // chart tab is missing so index is 1
-
-      this.drillDownComponent.SelectTab(this._drillDownSelectTab);
-      this._drillDownSelectTab = -1; // turn off flag
     }
   }
