+++ packages/Angular/Generic/skip-chat/src/public-api.ts
@@ -5,7 +5,7 @@ export * from './lib/module';
 export * from './lib/skip-chat/skip-chat.component';
 export * from './lib/skip-single-message/skip-single-message.component';
 export * from './lib/dynamic-report/skip-dynamic-report-wrapper';
-export * from './lib/dynamic-report/dynamic-report';
+export * from './lib/dynamic-report/linear-report';
 export * from './lib/dynamic-report/dynamic-chart';
 export * from './lib/dynamic-report/dynamic-grid';
-export * from './lib/dynamic-report/dynamic-drill-down';
+export * from './lib/drill-down-info';
